+++ public/modalidade/listModalidadeSearching.php
@@ -49,7 +49,7 @@ if ($dadosModalidade->num_rows > 0) {
                         <a href="../geracaoDeDocx/coletaModal.php?id=<?php echo $exibir["codigo_processo"] ?>" style="text-decoration: none; color: white" target="blank">Relatório</a>
                     </button>
                     <button type="submit" class="btn btn-danger btn-sm" formmethod="post">
-                        <a href="deleteModalidade.php?id=<?php echo $exibir["id_modalidade"] ?>" style="text-decoration: none; color: white"> Excluir </a>
+                        <a href="deleteModalidade.php?id=<?php echo $exibir["id_modalidade"] ?>" style="text-decoration: none; color: white" onclick="return confirm('Deseja mesmo apagar o processo?');"> Excluir </a>
                     </button>
                     <form name="editbutton" action="editModalidade.php" method="POST">
                         <input type="hidden" name="codigo" value="<?php echo $exibir["codigo_processo"] ?>">
